+++ php-templates/inertia.php
@@ -2,5 +2,9 @@
 
 echo json_encode([
     ...config('inertia.testing', []),
-    'page_paths' => collect(config('inertia.testing.page_paths', []))->map(fn($path) => LaravelVsCode::relativePath($path))->toArray(),
+    'page_paths' => collect(config('inertia.testing.page_paths', []))->flatMap(function($path) {
+        $relativePath = LaravelVsCode::relativePath($path);
+
+        return [$relativePath, mb_strtolower($relativePath)];
+    })->unique()->values(),
 ]);