+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the Laravel extension will be documented in this file.
 
 - Improve file watcher performance ([#327](https://github.com/laravel/vs-code-extension/pull/327))
 - Autocompletion, linking, and diagnostics for Storage Disks ([#328](https://github.com/laravel/vs-code-extension/pull/328))
+- Fix relative path when path does not exist ([#322](https://github.com/laravel/vs-code-extension/pull/322))
 
 ## [v1.0.3]